// apps/api-quickcart/src/modules/store/application/types/RegisterCustomer.types.ts
/**
 * Copyright (c) 2026 Ada Technology. All rights reserved.
 *
 * This source code is proprietary and confidential. Unauthorized copying,
 * modification, distribution, or use of this file, via any medium, is
 * strictly prohibited without prior written permission from Ada Technology.
 *
 * Author: Anderson Filho <andersonfrfilho@gmail.com>
 */

export type RegisterCustomerParams = {
  readonly name: string
  readonly email: string
  readonly phone: string
  readonly password: string
}

export type RegisterCustomerResult = {
  readonly userId: string
  readonly customerId: string
}